Partitioning Into Minimum Number Of Deci-Binary Numbers
A decimal number is called deci-binary if each of its digits is either 0 or 1 without any leading zeros. For example, 101 and 1100 are deci-binary, while 112 and 3001 are not.

Given a string n that represents a positive decimal integer, return the minimum number of positive deci-binary numbers needed so that they sum up to n.

 

Example 1:

Input: n = "32"
Output: 3
Explanation: 10 + 11 + 11 = 32

Example 2:

Input: n = "82734"
Output: 8

Example 3:

Input: n = "27346209830709182346"
Output: 9

 

Constraints:

    1 <= n.length <= 105
    n consists of only digits.
    n does not contain any leading zeros and represents a positive integer.

   Hide Hint #1  
Think about if the input was only one digit. Then you need to add up as many ones as the value of this digit.
   Hide Hint #2  
If the input has multiple digits, then you can solve for each digit independently, and merge the answers to form numbers that add up to that input.
   Hide Hint #3  
Thus the answer is equal to the max digit.

Idea:

If each deci-binary number has no higher than a 1 in each position, then it will take at least x numbers to achieve an x in any given position of n. This means that the largest character in any position in n will determine how many deci-binary numbers must be added together to obtain n.

We can quite easily separate the characters of n, find the max, and return that number.

    Time Complexity: O(N) where N is the length of the input string n
    Space Complexity: O(N) or O(1) depending on whether or not we split n to an array first