Maximum Area of a Piece of Cake After Horizontal and Vertical Cuts
Given a rectangular cake with height h and width w, and two arrays of integers horizontalCuts and verticalCuts where horizontalCuts[i] is the distance from the top of the rectangular cake to the ith horizontal cut and similarly, verticalCuts[j] is the distance from the left of the rectangular cake to the jth vertical cut.

Return the maximum area of a piece of cake after you cut at each horizontal and vertical position provided in the arrays horizontalCuts and verticalCuts. Since the answer can be a huge number, return this modulo 10^9 + 7.

 

Example 1:
https://assets.leetcode.com/uploads/2020/05/14/leetcode_max_area_2.png
Input: h = 5, w = 4, horizontalCuts = [1,2,4], verticalCuts = [1,3]
Output: 4 
Explanation: The figure above represents the given rectangular cake. Red lines are the horizontal and vertical cuts. After you cut the cake, the green piece of cake has the maximum area.

Example 2:
https://assets.leetcode.com/uploads/2020/05/14/leetcode_max_area_3.png
Input: h = 5, w = 4, horizontalCuts = [3,1], verticalCuts = [1]
Output: 6
Explanation: The figure above represents the given rectangular cake. Red lines are the horizontal and vertical cuts. After you cut the cake, the green and yellow pieces of cake have the maximum area.

Example 3:

Input: h = 5, w = 4, horizontalCuts = [3], verticalCuts = [3]
Output: 9

 

Constraints:

    2 <= h, w <= 10^9
    1 <= horizontalCuts.length < min(h, 10^5)
    1 <= verticalCuts.length < min(w, 10^5)
    1 <= horizontalCuts[i] < h
    1 <= verticalCuts[i] < w
    It is guaranteed that all elements in horizontalCuts are distinct.
    It is guaranteed that all elements in verticalCuts are distinct.

   Hide Hint #1  
Sort the arrays, then compute the maximum difference between two consecutive elements for horizontal cuts and vertical cuts.
   Hide Hint #2  
The answer is the product of these maximum values in horizontal cuts and vertical cuts.


Idea:

The trick to this problem is realizing that if the horizontal slices and vertical slices are perpendicular, then all vertical slices cross all horizontal slices. This means that we just need to find the largest of each, and the cross-section should be the largest slice.

To find the largest slice of each, we need to first sort the horizontal cuts (hc) and vertical cuts (vc), then iterate through both sets and keep track of the maximum difference found between two consecutive cuts (maxh, maxv). We need to not forget to include the two end cuts, which are found using 0 and h/w, as well.

Once we have the largest difference for both, we can just return the product of these two numbers, modulo 1e9+7.

    Time Complexity: O(N * log(N) + M * log(M)) where N is the length of hc and M is the length of vc
    Space Complexity: O(1)